// frontend/src/backend.d.ts
export class ExternalBlob {
    getBytes(): Promise<Uint8Array<ArrayBuffer>>;
    getDirectURL(): string;
    static fromURL(url: string): ExternalBlob;
    static fromBytes(blob: Uint8Array<ArrayBuffer>): ExternalBlob;
    withUploadProgress(onProgress: (percentage: number) => void): ExternalBlob;
}
export interface Load {
    weight: number;
    status: string;
    client: Principal;
    isApproved: boolean;
    description: string;
    truckType: TruckType;
    loadingLocation: string;
    tracking?: TrackingUpdate;
    assignedTransporter?: Principal;
    confirmation: LoadConfirmation;
    price: number;
    offloadingLocation: string;
}
// ...
export interface TrackingUpdate {
    status: string;
    timestamp: bigint;
    location: string;
}
export interface LoadConfirmation {
    orderId: string;
    confirmationFiles: Array<ExternalBlob>;
}
// ...
export enum ClientVerificationStatus {
    verified = "verified",
    pending = "pending",
    rejected = "rejected"
}
export enum TruckType {
    triaxle = "triaxle",
    superlinkFlatdeck = "superlinkFlatdeck",
    sideTipper = "sideTipper"
}